refactor(util): move LoadingBar from cli to util and update imports\n\nRelocate LoadingBar to util package for better organization and reuse.\nUpdate BookCommands to import com.penrose.bibby.util.LoadingBar.\n\nNo functional changes.
